fix hot-reload issue with DatContext by splitting react components from other exported functions
